double does make a difference, but don't solve all problems, using the w D2w > 0 cut alleviates some of the troubles, but not all. Also, there is some poor approximation of the weights on the border, that I have to check before continuing, the whole trouble was to get those right

+++ src/inputManager.h
@@ -4,10 +4,10 @@ typedef struct configVar {
   int Width,Height,nVortex,nRuns;
   int runType,genType,dim,adaptive;
   int numG,numRc,hNG,hNRc,hNa,hNb,hNN,hNmax;
-  double Gmin,Gmax,RcMin,RcMax;
-  double *Glist,*Rclist,x0[2],xf[2];
-  double swThresh, sndSwThresh,cutoff,v0y0;
-  double hGmin,hGmax,hRcMin,hRcMax;
+  float Gmin,Gmax,RcMin,RcMax;
+  float *Glist,*Rclist,x0[2],xf[2];
+  float swThresh, sndSwThresh,cutoff,v0y0;
+  float hGmin,hGmax,hRcMin,hRcMax;
   char *genFile,*tag,*folder;
 } configVar;
