+++ src/tests/test_addUSingleOseen3.c
@@ -10,9 +10,9 @@ int main(int argc,char **argv){
   const int Width = 100, Height = 100, Pop=10,nVortex=3;
   int i,j,err,ngbr,found,type=0;
   int nbList[8],label[Width*Height],eqList[Pop],**eqClass;
-  double parVortex[4*nVortex],x0[2],dx[2],xf[2],*sField=NULL;
-  double *gField=NULL,*uField=NULL,X[Width],Y[Height];
-  double x,y,v0y0 = 0.0;
+  float parVortex[4*nVortex],x0[2],dx[2],xf[2],*sField=NULL;
+  float *gField=NULL,*uField=NULL,X[Width],Y[Height];
+  float x,y,v0y0 = 0.0;
 
   eqClass=(int**)malloc(NumCls*sizeof(int*));
   if(eqClass==NULL)
@@ -30,7 +30,7 @@ int main(int argc,char **argv){
   parVortex[4+0]=1.; parVortex[4+1]=1.; parVortex[4+2]=2.; parVortex[4+3]=0.;
   parVortex[8+0]=1.; parVortex[8+1]=1.; parVortex[8+2]=0.; parVortex[8+3]=4.;
  
-  gField = (double *)malloc(4*Height*Width*sizeof(double));
+  gField = (float *)malloc(4*Height*Width*sizeof(float));
   if(gField==NULL){
     printf("memory not allocked\n");
     return 1;
@@ -38,7 +38,7 @@ int main(int argc,char **argv){
   for(i=0;i<4*Height*Width;i+=1)
     gField[i]=0.;
 
-  uField = (double *)malloc(2*Height*Width*sizeof(double));
+  uField = (float *)malloc(2*Height*Width*sizeof(float));
   if(uField==NULL){
     printf("memory not allocked\n");
     return 1;
